requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
